+++ tests/data/valid_data.py
@@ -5,6 +5,8 @@ VALID_CONTEXT = {
     "lightbulbColor": {"type": "enum"},
 }
 VALID_OUTPUT = ["lightbulbColor"]
+VALID_MODEL_TYPE = "boosting"
+VALID_TREE_MAX_DEPTH = 5
 VALID_TQ = 100
 VALID_TREE_MAX_OPERATIONS = 50000
 VALID_LEARNING_PERIOD = 6000000
@@ -12,6 +14,18 @@ VALID_CONFIGURATION = {
     "context": VALID_CONTEXT,
     "output": VALID_OUTPUT,
     "time_quantum": VALID_TQ,
+}
+VALID_NUM_ROUND = 100
+VALID_LEARNING_RATE = 0.1
+VALID_BOOSTING_CONFIGURATION = {
+    "model_type": VALID_MODEL_TYPE,
+    "context": VALID_CONTEXT,
+    "output": VALID_OUTPUT,
+    "num_iterations": VALID_NUM_ROUND,
+    "learning_rate": VALID_LEARNING_RATE,
+    "tree_max_depth": VALID_TREE_MAX_DEPTH,
+    "tree_max_operations": VALID_TREE_MAX_OPERATIONS,
+    "operations_as_events": True,
 }
 
 VALID_GENERATOR_CONFIGURATION = {
@@ -21,6 +35,17 @@ VALID_GENERATOR_CONFIGURATION = {
     "operations_as_events": True,
     "learning_period": VALID_LEARNING_PERIOD,
     "tree_max_operations": VALID_TREE_MAX_OPERATIONS,
+    "filter": ["test_filter"],
+}
+
+VALID_GENERATOR_BOOSTING_CONFIGURATION = {
+    "model_type": VALID_MODEL_TYPE,
+    "context": VALID_CONTEXT,
+    "output": VALID_OUTPUT,
+    "num_iterations": VALID_NUM_ROUND,
+    "learning_rate": VALID_LEARNING_RATE,
+    "tree_max_operations": VALID_TREE_MAX_OPERATIONS,
+    "operations_as_events": True,
     "filter": ["test_filter"],
 }